+++ LangtonMain.cpp
@@ -6,7 +6,7 @@
 #define MAX_WIDTH 100
 #define MAX_ITERATIONS 15000
 
-bool run_simulation(int indent);
+bool run_simulation(size_t indent);
 void print_board(const LangtonAnt& ant);
 
 int main(int argc, char* argv[]) {
@@ -25,17 +25,17 @@ int main(int argc, char* argv[]) {
  * @param indent the amount of spaces each prompt should be indented
  * @return true
  */
-bool run_simulation(int indent) {
-    int length = menu::prompt_range("board length", 1, MAX_LENGTH, indent);
-    int width = menu::prompt_range("board width", 1, MAX_WIDTH, indent);
-    int row = menu::prompt_range("starting row", 0, length - 1, indent);
-    int col = menu::prompt_range("starting column", 0, width - 1, indent);
+bool run_simulation(size_t indent) {
+    size_t length = menu::prompt_range("board length", 1, MAX_LENGTH, indent);
+    size_t width = menu::prompt_range("board width", 1, MAX_WIDTH, indent);
+    size_t row = menu::prompt_range("starting row", 0ul, length - 1, indent);
+    size_t col = menu::prompt_range("starting column", 0ul, width - 1, indent);
     const std::vector<std::string> options = {"NORTH", "EAST", "SOUTH", "WEST"};
-    int direction = menu::prompt_option("starting direction", options, indent) - 1;
-    int iterations = menu::prompt_range("number of iterations", 1, MAX_ITERATIONS, indent);
+    size_t direction = menu::prompt_option("starting direction", options, indent) - 1;
+    size_t iterations = menu::prompt_range("number of iterations", 1, MAX_ITERATIONS, indent);
 
     LangtonAnt ant = LangtonAnt(length, width, row, col, static_cast<Direction>(direction));
-    for (int i = 0; i < iterations; i++) {
+    for (size_t i = 0; i < iterations; i++) {
         ant.move();
     }
     print_board(ant);
@@ -51,8 +51,8 @@ bool run_simulation(int indent) {
  */
 void print_board(const LangtonAnt& ant) {
     auto board = ant.getBoard();
-    for (int i = 0; i < board.size(); i++) {
-        for (int j = 0; j < board[i].size(); j++) {
+    for (size_t i = 0; i < board.size(); i++) {
+        for (size_t j = 0; j < board[i].size(); j++) {
             if (i == ant.getRow() && j == ant.getCol()) {
                 std::cout << "*";
             } else if (board[i][j]) {
